test(update): isolate Bun ownership regression coverage
